Refactor application.js
