+++ app/assets/javascripts/application.js
@@ -2,50 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import Relay from "react-relay";
 
-class Plan extends React.Component {
-  render() {
-    var plan = this.props.plan;
-    return (
-      <div>
-        <span>{plan.name}</span><br />
-        <span>{plan.start}</span> - <span>{plan.finish}</span>
-      </div>
-    );
-  }
-}
-
-const PlanContainer = Relay.createContainer(Plan, {
-  fragments: {
-    plan: () => Relay.QL`
-      fragment on Plan {
-        name,
-        start,
-        finish
-      }
-    `
-  }
-});
-
-class Calendar extends React.Component {
-  render() {
-    var calendar = this.props.calendar;
-    return (
-      <PlanContainer plan={calendar.plan} />
-    );
-  }
-}
-
-const CalendarContainer =  Relay.createContainer(Calendar, {
-  fragments: {
-    calendar: () => Relay.QL`
-      fragment on Calendar {
-        plan {
-          ${PlanContainer.getFragment('plan')}
-        }
-      }
-    `
-  }
-});
+import Calendar from "./components/Calendar";
 
 class CalendarRoute extends Relay.Route {
   static queries = {
@@ -61,13 +18,13 @@ class CalendarRoute extends Relay.Route {
   static routeName = "Calendar";
 }
 
-const calendarRoute = new CalendarRoute({date: "2017-3-15"});
+const calendarRoute = new CalendarRoute({date: "2017-3-14"});
 
 window.onload = function() {
   const calendarAppContainer = document.getElementById("calendar_app");
   if (calendarAppContainer) {
     ReactDOM.render(
-      <Relay.RootContainer Component={CalendarContainer} route={calendarRoute} />,
+      <Relay.RootContainer Component={Calendar} route={calendarRoute} />,
       calendarAppContainer
     );
   }
